+++ @xen-orchestra/lite/src/views/host/HostNetworkView.vue
@@ -1,10 +1,10 @@
 <template>
-  <div class="host-network-view">
+  <div class="host-network-view" :class="{ mobile: uiStore.isMobile }">
     <UiCard class="container">
       <HostPifsTable :pifs />
     </UiCard>
-    <HostPifSidePanel v-if="selectedPif" :pif="selectedPif" />
-    <UiPanel v-else>
+    <HostPifSidePanel v-if="selectedPif" :pif="selectedPif" @close="selectedPif = undefined" />
+    <UiPanel v-else-if="!uiStore.isMobile">
       <VtsNoSelectionHero type="panel" />
     </UiPanel>
   </div>
@@ -21,16 +21,19 @@ import VtsNoSelectionHero from '@core/components/state-hero/VtsNoSelectionHero.v
 import UiCard from '@core/components/ui/card/UiCard.vue'
 import UiPanel from '@core/components/ui/panel/UiPanel.vue'
 import { useRouteQuery } from '@core/composables/route-query.composable'
+import { useUiStore } from '@core/stores/ui.store.ts'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useRoute } from 'vue-router'
 
 const { records } = usePifStore().subscribe()
 const { getByOpaqueRef: getHostOpaqueRef } = useHostStore().subscribe()
+const uiStore = useUiStore()
 
 const route = useRoute()
 
-usePageTitleStore().setTitle(useI18n().t('network'))
+const { t } = useI18n()
+usePageTitleStore().setTitle(t('network'))
 
 const pifs = computed(() => {
   return records.value.filter(pif => {
@@ -48,13 +51,15 @@ const selectedPif = useRouteQuery<XenApiPif | undefined>('id', {
 
 <style lang="postcss" scoped>
 .host-network-view {
-  display: grid;
-  grid-template-columns: minmax(0, 1fr) 40rem;
+  &:not(.mobile) {
+    display: grid;
+    grid-template-columns: minmax(0, 1fr) 40rem;
+  }
 
   .container {
     height: fit-content;
-    gap: 4rem;
     margin: 0.8rem;
+    gap: 4rem;
   }
 }
 </style>