+++ @xen-orchestra/web/src/components/pool/network/PoolNetworkSidePanel.vue
@@ -1,27 +1,40 @@
 <template>
-  <UiPanel>
+  <UiPanel :class="{ 'mobile-drawer': uiStore.isMobile }">
     <template #header>
-      <UiButton
-        v-tooltip="t('coming-soon')"
-        disabled
-        variant="tertiary"
-        size="medium"
-        accent="brand"
-        :left-icon="faEdit"
-      >
-        {{ t('edit') }}
-      </UiButton>
-      <UiButton
-        v-tooltip="t('coming-soon')"
-        disabled
-        variant="tertiary"
-        size="medium"
-        accent="danger"
-        :left-icon="faTrash"
-      >
-        {{ t('delete') }}
-      </UiButton>
-      <UiButtonIcon v-tooltip="t('coming-soon')" disabled accent="brand" size="medium" :icon="faEllipsis" />
+      <div :class="{ 'action-buttons-container': uiStore.isMobile }">
+        <UiButtonIcon
+          v-if="uiStore.isMobile"
+          v-tooltip="t('close')"
+          size="medium"
+          variant="tertiary"
+          accent="brand"
+          :icon="faAngleLeft"
+          @click="emit('close')"
+        />
+        <div class="action-buttons">
+          <UiButton
+            v-tooltip="t('coming-soon')"
+            disabled
+            size="medium"
+            variant="tertiary"
+            accent="brand"
+            :left-icon="faEdit"
+          >
+            {{ t('edit') }}
+          </UiButton>
+          <UiButton
+            v-tooltip="t('coming-soon')"
+            disabled
+            size="medium"
+            variant="tertiary"
+            accent="danger"
+            :left-icon="faTrash"
+          >
+            {{ t('delete') }}
+          </UiButton>
+          <UiButtonIcon v-tooltip="t('coming-soon')" disabled accent="brand" size="medium" :icon="faEllipsis" />
+        </div>
+      </div>
     </template>
     <template #default>
       <UiCard class="card-container">
@@ -52,10 +65,10 @@
             </template>
           </VtsCardRowKeyValue>
           <!-- VLAN -->
-          <VtsCardRowKeyValue>
+          <VtsCardRowKeyValue v-if="networkVlan">
             <template #key>{{ t('vlan') }}</template>
             <template #value>{{ networkVlan }}</template>
-            <template v-if="pifs[0].vlan !== -1" #addons>
+            <template #addons>
               <VtsCopyButton :value="String(networkVlan)" />
             </template>
           </VtsCardRowKeyValue>
@@ -117,8 +130,8 @@
 
 <script setup lang="ts">
 import PifRow from '@/components/pif/PifRow.vue'
-import { usePifStore } from '@/stores/xo-rest-api/pif.store'
-import type { XoNetwork } from '@/types/xo/network.type'
+import { usePifStore } from '@/stores/xo-rest-api/pif.store.ts'
+import type { XoNetwork } from '@/types/xo/network.type.ts'
 import VtsCardRowKeyValue from '@core/components/card/VtsCardRowKeyValue.vue'
 import VtsCopyButton from '@core/components/copy-button/VtsCopyButton.vue'
 import UiButton from '@core/components/ui/button/UiButton.vue'
@@ -127,8 +140,9 @@ import UiCard from '@core/components/ui/card/UiCard.vue'
 import UiCardTitle from '@core/components/ui/card-title/UiCardTitle.vue'
 import UiCounter from '@core/components/ui/counter/UiCounter.vue'
 import UiPanel from '@core/components/ui/panel/UiPanel.vue'
-import { vTooltip } from '@core/directives/tooltip.directive'
-import { faEdit, faEllipsis, faTrash } from '@fortawesome/free-solid-svg-icons'
+import { vTooltip } from '@core/directives/tooltip.directive.ts'
+import { useUiStore } from '@core/stores/ui.store.ts'
+import { faAngleLeft, faEdit, faEllipsis, faTrash } from '@fortawesome/free-solid-svg-icons'
 import { computed } from 'vue'
 import { useI18n } from 'vue-i18n'
 
@@ -136,7 +150,12 @@ const { network } = defineProps<{
   network: XoNetwork
 }>()
 
+const emit = defineEmits<{
+  close: []
+}>()
+
 const { getPifsByNetworkId } = usePifStore().subscribe()
+const uiStore = useUiStore()
 
 const { t } = useI18n()
 
@@ -144,8 +163,9 @@ const pifs = computed(() => getPifsByNetworkId(network.id))
 
 const networkVlan = computed(() => {
   if (pifs.value.length === 0) {
-    return ''
+    return
   }
+
   return pifs.value[0].vlan !== -1 ? pifs.value[0].vlan.toString() : t('none')
 })
 
@@ -185,6 +205,23 @@ const pifsCount = computed(() => pifs.value.length)
     thead tr th {
       color: var(--color-neutral-txt-secondary);
     }
+  }
+}
+
+.mobile-drawer {
+  position: fixed;
+  inset: 0;
+
+  .action-buttons-container {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+  }
+
+  .action-buttons {
+    display: flex;
+    align-items: center;
   }
 }
 </style>