+++ @xen-orchestra/web/src/components/pool/network/PoolNetworksTable.vue
@@ -1,18 +1,11 @@
 <template>
-  <div class="pool-host-internal-networks-table">
+  <div class="pool-networks-table">
     <UiTitle>
-      {{ t('host-internal-networks') }}
+      {{ t('networks') }}
       <template #actions>
-        <UiButton
-          v-tooltip="t('coming-soon')"
-          disabled
-          :left-icon="faPlus"
-          variant="secondary"
-          accent="brand"
-          size="medium"
-        >
+        <UiDropdownButton v-tooltip="t('coming-soon')" disabled>
           {{ t('new') }}
-        </UiButton>
+        </UiDropdownButton>
       </template>
     </UiTitle>
     <div class="container">
@@ -28,6 +21,16 @@
             size="medium"
           >
             {{ t('edit') }}
+          </UiButton>
+          <UiButton
+            v-tooltip="t('coming-soon')"
+            disabled
+            :left-icon="faCopy"
+            variant="tertiary"
+            accent="brand"
+            size="medium"
+          >
+            {{ t('copy-info-json') }}
           </UiButton>
           <UiButton
             v-tooltip="t('coming-soon')"
@@ -91,6 +94,7 @@
                 disabled
                 size="small"
               />
+              <VtsConnectionStatus v-else-if="column.id === 'status'" :status="column.value" />
               <div v-else v-tooltip="{ placement: 'bottom-end' }" class="text-ellipsis">
                 {{ column.value }}
               </div>
@@ -107,30 +111,34 @@
 </template>
 
 <script setup lang="ts">
-import { useNetworkStore } from '@/stores/xo-rest-api/network.store'
-import type { XoNetwork } from '@/types/xo/network.type'
+import { useNetworkStore } from '@/stores/xo-rest-api/network.store.ts'
+import { usePifStore } from '@/stores/xo-rest-api/pif.store.ts'
+import type { XoNetwork } from '@/types/xo/network.type.ts'
+import VtsConnectionStatus from '@core/components/connection-status/VtsConnectionStatus.vue'
 import VtsDataTable from '@core/components/data-table/VtsDataTable.vue'
 import VtsIcon from '@core/components/icon/VtsIcon.vue'
 import VtsStateHero from '@core/components/state-hero/VtsStateHero.vue'
 import UiButton from '@core/components/ui/button/UiButton.vue'
 import UiButtonIcon from '@core/components/ui/button-icon/UiButtonIcon.vue'
 import UiCheckbox from '@core/components/ui/checkbox/UiCheckbox.vue'
+import UiDropdownButton from '@core/components/ui/dropdown-button/UiDropdownButton.vue'
 import UiQuerySearchBar from '@core/components/ui/query-search-bar/UiQuerySearchBar.vue'
 import UiTableActions from '@core/components/ui/table-actions/UiTableActions.vue'
 import UiTitle from '@core/components/ui/title/UiTitle.vue'
 import UiTopBottomTable from '@core/components/ui/top-bottom-table/UiTopBottomTable.vue'
-import { useRouteQuery } from '@core/composables/route-query.composable'
-import useMultiSelect from '@core/composables/table/multi-select.composable'
-import { useTable } from '@core/composables/table.composable'
-import { vTooltip } from '@core/directives/tooltip.directive'
+import { useRouteQuery } from '@core/composables/route-query.composable.ts'
+import useMultiSelect from '@core/composables/table/multi-select.composable.ts'
+import { useTable } from '@core/composables/table.composable.ts'
+import { vTooltip } from '@core/directives/tooltip.directive.ts'
 import type { IconDefinition } from '@fortawesome/fontawesome-common-types'
 import {
   faAlignLeft,
   faCaretDown,
+  faCopy,
   faEdit,
   faEllipsis,
   faHashtag,
-  faPlus,
+  faPowerOff,
   faTrash,
 } from '@fortawesome/free-solid-svg-icons'
 import { noop } from '@vueuse/shared'
@@ -141,11 +149,15 @@ const { networks } = defineProps<{
   networks: XoNetwork[]
 }>()
 
+const { t } = useI18n()
+
 const { isReady, hasError } = useNetworkStore().subscribe()
 
-const { t } = useI18n()
-const searchQuery = ref('')
+const { records: pifs } = usePifStore().subscribe()
+
 const selectedNetworkId = useRouteQuery('id')
+
+const searchQuery = ref('')
 
 const filteredNetworks = computed(() => {
   const searchTerm = searchQuery.value.trim().toLocaleLowerCase()
@@ -167,6 +179,32 @@ const toggleSelect = () => {
   selected.value = selected.value.length === 0 ? networkIds.value : []
 }
 
+const getNetworkVlan = (network: XoNetwork) => {
+  const networkPIFs = pifs.value.filter(pif => network.PIFs.includes(pif.id))
+
+  if (networkPIFs.length > 0) {
+    return networkPIFs[0].vlan !== -1 ? networkPIFs[0].vlan.toString() : t('none')
+  }
+}
+
+const getNetworkStatus = (network: XoNetwork) => {
+  const networkPIFs = pifs.value.filter(pif => network.PIFs?.includes(pif.id))
+
+  if (networkPIFs.length === 0) {
+    return 'disconnected'
+  }
+
+  const isConnected = networkPIFs.map(pif => pif.attached && pif.carrier)
+  if (isConnected.every(Boolean)) {
+    return 'connected'
+  }
+
+  if (isConnected.some(Boolean)) {
+    return 'partially-connected'
+  }
+  return 'disconnected'
+}
+
 const getLockingMode = (lockingMode: string) => (lockingMode === 'disabled' ? t('disabled') : t('unlocked'))
 
 const { visibleColumns, rows } = useTable('networks', filteredNetworks, {
@@ -174,7 +212,11 @@ const { visibleColumns, rows } = useTable('networks', filteredNetworks, {
   columns: define => [
     define('checkbox', noop, { label: '', isHideable: false }),
     define('name_label', { label: t('name') }),
-    define('name_description', { label: t('description') }),
+    define('name_description', {
+      label: t('description'),
+    }),
+    define('status', record => getNetworkStatus(record), { label: t('pifs-status') }),
+    define('vlan', record => getNetworkVlan(record), { label: t('vlan') }),
     define('MTU', { label: t('mtu') }),
     define('default_locking_mode', record => getLockingMode(record.default_locking_mode), {
       label: t('default-locking-mode'),
@@ -183,25 +225,27 @@ const { visibleColumns, rows } = useTable('networks', filteredNetworks, {
   ],
 })
 
-type NetworkHeader = 'name_label' | 'name_description' | 'MTU' | 'default_locking_mode'
+type NetworkHeader = 'name_label' | 'name_description' | 'status' | 'vlan' | 'MTU' | 'default_locking_mode'
 
 const headerIcon: Record<NetworkHeader, IconDefinition> = {
   name_label: faAlignLeft,
   name_description: faAlignLeft,
+  status: faPowerOff,
+  vlan: faAlignLeft,
   MTU: faHashtag,
   default_locking_mode: faCaretDown,
 }
 </script>
 
 <style scoped lang="postcss">
-.pool-host-internal-networks-table,
+.pool-networks-table,
 .table-actions,
 .container {
   display: flex;
   flex-direction: column;
 }
 
-.pool-host-internal-networks-table {
+.pool-networks-table {
   gap: 2.4rem;
 
   .container,